+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/model/RoomSummary.kt
@@ -42,7 +42,8 @@ data class RoomSummary(
         val versioningState: VersioningState = VersioningState.NONE,
         val readMarkerId: String? = null,
         val userDrafts: List<UserDraft> = emptyList(),
-        var isEncrypted: Boolean
+        var isEncrypted: Boolean,
+        val typingRoomMemberIds: List<String> = emptyList()
 ) {
 
     val isVersioned: Boolean